pyproject.toml
[project]
name = "aty"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-auth>=2.49.2",
    "gspread>=6.2.1",
    "python-dotenv>=1.2.2",
    "questionary>=2.1.1",
    "typer>=0.24.1",
]
